+++ framework_src/FreSharp/FreSharpHelper.cs
@@ -36,6 +36,7 @@ namespace TuaRua.FreSharp {
         /// <param name="value"></param>
         /// <returns></returns>
         public static FreObjectSharp FreObjectSharpFromObject(object value) {
+            if (value == null) return null;
             var t = value.GetType();
             if (t == typeof(FREObject)) {
                 return new FreObjectSharp((FREObject) value);
@@ -194,7 +195,7 @@ namespace TuaRua.FreSharp {
         /// </summary>
         /// <param name="rawValue"></param>
         /// <returns></returns>
-        public static DateTime GetAsDateTime(FREObject rawValue) => 
+        public static DateTime GetAsDateTime(FREObject rawValue) =>
             new DateTime(1970, 1, 1, 0, 0, 0, 0).AddSeconds(rawValue.GetProp("time").AsDouble() / 1000);